25-11-clearing-relations

+++ 25 Rest Application/backend/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -50,20 +51,30 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path.replace(/\\/g, '/');
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: {
-            name: 'Maximilian',
-        },
+        creator: req.userId,
     });
     post.save()
+        .then(() => {
+            return User.findById(req.userId);
+        })
+        .then((user) => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
         .then((result) => {
-            console.log(result);
             res.status(201).json({
                 message: 'Post created successfully!',
                 post: post,
+                creator: {
+                    _id: creator._id,
+                    name: creator.name,
+                },
             });
         })
         .catch((err) => {
@@ -125,6 +136,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized.');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -153,13 +169,24 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized.');
+                error.statusCode = 403;
+                throw error;
+            }
             // Check logged in user
             clearImage(post.imageUrl);
             /** .findByIdAndRemove -- REMOVED */
             return Post.findByIdAndDelete(postId);
         })
         .then((result) => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then((user) => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then((result) => {
             res.status(200).json({ message: 'Deleted post.' });
         })
         .catch((err) => {
